refactor: Member 서비스 @Builder 추가
