+++ src/main/java/com/mjuAppSW/joA/geography/location/LocationJpaRepository.java
